code fixes and refactoring based on pull request comments
