--- functions/.env.example.json
@@ -1,8 +1,0 @@
-{
-  "testingConfig": {
-    "databaseURL": "",
-    "storageBucket": "",
-    "projectId": "",
-    "serviceAccountKey": ""
-  }
-}
