+++ functions/env.example.json
@@ -1,0 +1,6 @@
+{
+  "databaseURL": "",
+  "storageBucket": "",
+  "projectId": "",
+  "serviceAccountKey": ""
+}
