+++ functions/test/db/users/onDelete.test.ts
@@ -1,7 +1,7 @@
-import { admin, test } from '../../util/admin';
+import { admin, test } from '../../util/config';
 import * as chai from 'chai';
 import * as chaiAsPromised from 'chai-as-promised';
-import * as onDelete from '../../../src/db/users/onDelete.function';
+import onDelete from '../../../src/db/users/onDelete.function';
 import 'mocha';
 
 chai.use(chaiAsPromised);
@@ -9,32 +9,24 @@ chai.use(chaiAsPromised);
 describe('onDelete', () => {
   let userRecord: any;
 
-  before(async () => {
-    const user = {
-      uid: '1234',
-      email: 'user@example.com',
-      password: 'secretPassword'
-    };
-    userRecord = await admin.auth().createUser(user);
-  });
+  it('should delete the user from the authentication section', async () => {
+    userRecord = await admin.auth().createUser({ email: 'user@example.com' });
 
-  it('should delete the user from the authentication section', () => {
-    const wrapped = test.wrap(onDelete.default);
+    const wrapped = test.wrap(onDelete);
 
-    return wrapped(
+    await wrapped(
       {},
       {
         params: {
           uid: userRecord.uid
         }
       }
-    ).then(async () => {
-      await chai
-        .expect(admin.auth().getUser(userRecord.uid))
-        .to.be.rejectedWith(
-          Error,
-          'There is no user record corresponding to the provided identifier.'
-        );
-    });
+    );
+    return await chai
+      .expect(admin.auth().getUser(userRecord.uid))
+      .to.be.rejectedWith(
+        Error,
+        'There is no user record corresponding to the provided identifier.'
+      );
   });
 });
