+++ functions/test/https/createUser.test.ts
@@ -1,8 +1,11 @@
-import { admin, test } from '../util/admin';
+import { admin, test } from '../util/config';
 import { https } from 'firebase-functions';
 import * as chai from 'chai';
+import * as chaiAsPromised from 'chai-as-promised';
 import * as createUser from '../../src/https/createUser.function';
 import 'mocha';
+
+chai.use(chaiAsPromised);
 
 describe('createUser', () => {
   let userRecord: any;
